+++ NEWS.md
@@ -2,6 +2,7 @@
   - Copying OTP to clipboard and typing it are no longer exclusive
   - Deprecated copying to clipboard by default in favour of specifying
     `--clipboard` explicitly
+  - New option `--stdout` to print OTP to standard output
 
 - Version 0.11.0 - released 2020-08-27
   - `--clipboard` option removed in favour of `--clipboard-cmd`